+++ tools/tasks/customizer.js
@@ -5,14 +5,11 @@ var path = require('path');
 var fs = require('fs');
 var _ = require('lodash');
 var format = require('util').format;
-var browserify = require('browserify');
-var source = require('vinyl-source-stream');
-var buffer = require('vinyl-buffer');
+var webpack = require('webpack-stream');
 var del = require('del');
 var runSequence = require('run-sequence');
 var gulp = require('gulp');
 var $ = require('gulp-load-plugins')();
-var collapser = require('bundle-collapser/plugin');
 var file = require('../lib/file');
 var cstmzPath = path.join(__dirname, '../../dist/customized');
 var cstmzTmp = path.join(__dirname, '../../.cstmz-tmp');
@@ -127,12 +124,30 @@ gulp.task('customizer:less', function() {
 });
 
 gulp.task('customizer:js', function() {
-  return browserify({
-    entries: DEFAULTS.js,
-    paths: [path.join(__dirname, '../../js'), path.join(__dirname, '../../widget')]
-  }).plugin(collapser).bundle()
-    .pipe(source('amazeui.custom.js'))
-    .pipe(buffer())
+  return gulp.src(DEFAULTS.js)
+    .pipe(webpack({
+      output: {
+        filename: 'amazeui.custom.js',
+        library: 'AMUI',
+        libraryTarget: 'umd'
+      },
+      externals: [
+        {
+          jquery: {
+            root: 'jQuery',
+            commonjs2: 'jquery',
+            commonjs: 'jquery',
+            amd: 'jquery'
+          }
+        }
+      ],
+      resolve: {
+        modulesDirectories: [
+          path.join(__dirname, '../../js'),
+          path.join(__dirname, '../../widget')
+        ]
+      }
+    }))
     .pipe(gulp.dest(cstmzPath))
     .pipe($.uglify({
       output: {
@@ -145,8 +160,8 @@ gulp.task('customizer:js', function() {
     .pipe($.size({showFiles: true, gzip: true, title: 'gzipped'}));
 });
 
-gulp.task('customizer:clean', function(cb) {
-  del(DEFAULTS.tmp, cb);
+gulp.task('customizer:clean', function() {
+  return del(DEFAULTS.tmp);
 });
 
 gulp.task('customizer', function(cb) {